refactor(runner): split executor test support harness
