+++ crates/runner/src/executor/tests/agent_run_tests.rs
@@ -1,4 +1,18 @@
-use super::*;
+use std::sync::Arc;
+use std::time::Duration;
+
+use sandbox::{ProcessExit, ProcessOutputChunk, SandboxConfig, SandboxFactory, SandboxId};
+use sandbox_mock::MockSandboxFactory;
+
+use super::super::agent_run::{ProcessCancelTimeouts, RunStart, run_in_sandbox};
+use super::super::diagnostics::AgentStdoutStreamDiagnostics;
+use super::super::{EXIT_SIGKILL, PROCESS_CANCEL_WRITE_TIMEOUT};
+use super::support::{
+    CancelAfterWaitSandbox, RUN_IN_SANDBOX_TEST_TIMEOUT, create_overridden_sandbox,
+    minimal_context, spawn_run_in_sandbox_test, spawn_run_in_sandbox_test_with_timeouts,
+    test_executor_config, test_telemetry,
+};
+use crate::types::SandboxReuseResult;
 
 #[tokio::test]
 async fn run_in_sandbox_preserves_wait_result_when_cancel_arrives_after_wait() {
