+++ crates/runner/src/executor/tests/support/tracing.rs
@@ -1,0 +1,65 @@
+use std::collections::BTreeMap;
+use std::fmt;
+use std::sync::{Arc, Mutex};
+
+use tracing::field::{Field, Visit};
+use tracing::{Event, Level, Subscriber};
+use tracing_subscriber::layer::{Context, Layer};
+
+#[derive(Clone, Debug)]
+pub(in crate::executor::tests) struct CapturedEvent {
+    pub(in crate::executor::tests) level: Level,
+    pub(in crate::executor::tests) fields: BTreeMap<String, String>,
+}
+
+#[derive(Clone, Default)]
+pub(in crate::executor::tests) struct CapturedEvents {
+    events: Arc<Mutex<Vec<CapturedEvent>>>,
+}
+
+impl CapturedEvents {
+    pub(in crate::executor::tests) fn entries(&self) -> Vec<CapturedEvent> {
+        self.events.lock().unwrap().clone()
+    }
+}
+
+impl<S> Layer<S> for CapturedEvents
+where
+    S: Subscriber,
+{
+    fn on_event(&self, event: &Event<'_>, _ctx: Context<'_, S>) {
+        let mut visitor = CapturedFields::default();
+        event.record(&mut visitor);
+        self.events.lock().unwrap().push(CapturedEvent {
+            level: *event.metadata().level(),
+            fields: visitor.fields,
+        });
+    }
+}
+
+#[derive(Default)]
+struct CapturedFields {
+    fields: BTreeMap<String, String>,
+}
+
+impl Visit for CapturedFields {
+    fn record_str(&mut self, field: &Field, value: &str) {
+        self.fields
+            .insert(field.name().to_string(), value.to_string());
+    }
+
+    fn record_i64(&mut self, field: &Field, value: i64) {
+        self.fields
+            .insert(field.name().to_string(), value.to_string());
+    }
+
+    fn record_u64(&mut self, field: &Field, value: u64) {
+        self.fields
+            .insert(field.name().to_string(), value.to_string());
+    }
+
+    fn record_debug(&mut self, field: &Field, value: &dyn fmt::Debug) {
+        self.fields
+            .insert(field.name().to_string(), format!("{value:?}"));
+    }
+}
