+++ crates/runner/src/executor/tests/telemetry_tests.rs
@@ -1,4 +1,17 @@
-use super::*;
+use std::time::Duration;
+
+use sandbox::SandboxId;
+use sandbox_mock::MockSandboxFactory;
+
+use super::super::telemetry::{elapsed_since_api_start_ms, record_reuse_result};
+use super::super::{NewSandboxDispatch, execute_job, execute_job_reuse};
+use super::support::{
+    default_params, make_reusable_idle_sandbox, minimal_context, test_executor_config,
+};
+use crate::http::{HttpClient, HttpClientConfig};
+use crate::ids::RunId;
+use crate::telemetry::JobTelemetry;
+use crate::types::SandboxReuseResult;
 
 #[test]
 fn elapsed_since_api_start_ms_returns_elapsed_duration() {
